# net: gptp: fix follow_up timestamp and announce byte order