--- a/subsys/net/l2/ethernet/gptp/gptp_md.c
+++ b/subsys/net/l2/ethernet/gptp/gptp_md.c
@@ -61,13 +61,6 @@ static void gptp_md_follow_up_prepare(struct net_pkt *pkt,
 
 	hdr->log_msg_interval = sync_send->log_msg_interval;
 
-	fup->prec_orig_ts_secs_high =
-		htons(sync_send->precise_orig_ts._sec.high);
-	fup->prec_orig_ts_secs_low =
-		htonl(sync_send->precise_orig_ts._sec.low);
-	fup->prec_orig_ts_nsecs =
-		htonl(sync_send->precise_orig_ts.nanosecond);
-
 	fup->tlv_hdr.type = htons(GPTP_TLV_ORGANIZATION_EXT);
 	fup->tlv_hdr.len = htons(sizeof(struct gptp_follow_up_tlv));
 	fup->tlv.org_id[0] = GPTP_FUP_TLV_ORG_ID_BYTE_0;