This PR include gPTP accuracy optimize and some byte order fix.

1. Change follow_up message preciseOriginTimestamp from software trigger to hardware egress capture
2. Fix announce message `cur_utc_offset` and `offset_scaled_log_var` fields byte order.

Using mimxrt1050-evk board and Intel i210 ethernet card direct cable connect, host using linuxptp stack,
and test PPS phase between 1588_EVENT2_OUT and SDP1 by oscilloscope.

Before:
mimxrt1050-evk as grandmaster, i210 as slave, PPS offset is 3.5ms
i210 as grandmaster, mimxrt1050-evk as slave, PPS offset is 20~30us

After:
mimxrt1050-evk as grandmaster, i210 as slave, PPS offset is 540~580ns
i210 as grandmaster, mimxrt1050-evk as slave, PPS offset is 1250~1380ns