@chencaidy 2023-07-19
> @chencaidy, thank you for the contribution!
> 
> @hakehuang Can you test this PR to see how it does against #33747?

Here is my test report, may be able to offer some help.

OS: Ubuntu 22.04 with 5.19 kernel
Linuxptp version: [aa60db2](https://github.com/richardcochran/linuxptp/commit/aa60db270be12e7144e7b0cef2f5dde4213b7057)
testptp: https://elixir.bootlin.com/linux/v5.19/source/tools/testing/selftests/ptp/testptp.c

1. Build linuxptp and testptp (gcc -o testptp testptp.c)
2. Modity linuxptp/configs/gPTP.conf, `neighborPropDelayThresh 50000`, because mcux driver issue make path delay too high
3. Build gptp sample: west build -p always -b mimxrt1050_evk samples/net/gptp
4. If using mimxrt1050_evk as grandmaster, set `CONFIG_NET_GPTP_GM_CAPABLE=y` in prj.conf. using i120 as grandmaster, set to n, then flash to board.
5. Start host linuxptp: sudo ./ptp4l -f configs/gPTP.cfg -i `enp1s0` -m
6. Enable pps out in SDP1: sudo ./testptp -d /dev/ptp0 -L 1,2 -p 1000000000
If SDP1 pps out missing, rerun step 6.

Here is my test platform:
![platform](https://github.com/zephyrproject-rtos/zephyr/assets/18113342/3adf49e9-278a-4cac-9792-891c9cd6b229)

- rt1052 as grandmaster before

![linuxptp](https://github.com/zephyrproject-rtos/zephyr/assets/18113342/74ab1647-362a-4aac-9c4a-5b5f389c203c)
linuxptp: path delay >18000ns, rms > 2000ns, sometimes have a jerk.

![oscilloscope](https://github.com/zephyrproject-rtos/zephyr/assets/18113342/cf98569c-c24e-4417-b11a-2cb612f6e844)

- rt1052 as grandmaster after

![linuxptp](https://github.com/zephyrproject-rtos/zephyr/assets/18113342/4f4392ad-a74e-4b19-a31a-337c472665eb)
linuxptp: delay < 900ns, rms < 50ns, no jerk.

![oscilloscope](https://github.com/zephyrproject-rtos/zephyr/assets/18113342/c557ddaf-e407-41a2-a4f6-6b103012e856)

- host i210 as grandmaster before

![oscilloscope](https://github.com/zephyrproject-rtos/zephyr/assets/18113342/f9c7050c-1717-44a8-8a81-dd12686432ce)

- host i210 as grandmaster after

![oscilloscope](https://github.com/zephyrproject-rtos/zephyr/assets/18113342/5c195236-6331-4be7-a730-3efc1fd8ac41)