@hakehuang 2023-07-27
@chencaidy please take a looks at https://github.com/zephyrproject-rtos/zephyr/pull/60855 which should fix the jitter issue, and sync the timestamp in 24 hours